nustar_tools/requirements.txt
astropy==5.1.1
imageio==2.22.4
matplotlib==3.5.1
numpy==1.21.5
nustar_pysolar.egg==info
parse==1.19.0
photutils==1.5.0
ruptures==1.1.7
scikit_image==0.19.3
scipy==1.8.0
skimage==0.0
sunpy==4.1.0
